_app/requirements.txt
-i https://pypi.python.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
async-timeout==4.0.3 ; python_full_version <= '3.11.2'
azure-core==1.29.5
azure-storage-blob==12.19.0
brotli==1.1.0
certifi==2023.11.17 ; python_version >= '3.6'
cffi==1.16.0 ; python_version >= '3.8'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
cryptography==41.0.5 ; python_version >= '3.7'
django==4.2.7
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-rest-knox==4.2.0
django-storages[azure]==1.14.2
djangorestframework==3.14.0
gunicorn==21.2.0
hiredis==2.2.3
idna==3.4 ; python_version >= '3.5'
isodate==0.6.1
packaging==23.2 ; python_version >= '3.7'
pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4 ; python_version >= '3.5'
typing-extensions==4.8.0 ; python_version < '3.11'
urllib3==2.1.0 ; python_version >= '3.8'
whitenoise[brotli]==6.6.0
